+++ src/invader.cmake
@@ -18,6 +18,7 @@ set(INVADER_PARSER_FILES
     "${CMAKE_CURRENT_BINARY_DIR}/parser-compare.cpp"
     "${CMAKE_CURRENT_BINARY_DIR}/parser-normalize.cpp"
     "${CMAKE_CURRENT_BINARY_DIR}/parser-read-hek-file.cpp"
+    "${CMAKE_CURRENT_BINARY_DIR}/parser-check-uppercase-references.cpp"
     "${CMAKE_CURRENT_BINARY_DIR}/bitfield.cpp"
     "${CMAKE_CURRENT_BINARY_DIR}/enum.cpp"
 )
@@ -64,6 +65,7 @@ set(INVADER_SOURCE_FILES
     "${CMAKE_CURRENT_BINARY_DIR}/parser-compare.cpp"
     "${CMAKE_CURRENT_BINARY_DIR}/parser-normalize.cpp"
     "${CMAKE_CURRENT_BINARY_DIR}/parser-read-hek-file.cpp"
+    "${CMAKE_CURRENT_BINARY_DIR}/parser-check-uppercase-references.cpp"
     "${CMAKE_CURRENT_BINARY_DIR}/bitfield.cpp"
     "${CMAKE_CURRENT_BINARY_DIR}/enum.cpp"